[#11] Split AST types into different modules
Resolves #11
